requirements.txt
streamlit
requests
pytest
python-dotenv
pyodbc